Refactorign:  move some classes, and rename some classes.
