+++ JChat/Category/NSString+MessageInputView.h
