+++ JChat/Category/NSString+MessageInputView.m
