+++ JChat/Category/UIImage +ResizeMagick.m
