+++ JChat/Category/UIImage+ResizeMagick.h
